git-svn-id: https://svn.code.sf.net/p/oorexx/code-0/interpreter-3.x/trunk@1123 0b6cbdbe-3aab-466e-b73a-abd511dda0a2

+++ kernel/platform/unix/ExternalFunctions.cpp
@@ -839,7 +839,7 @@ void RestoreEnvironment(
   }
                                        /* Loop through the saved env */
                                        /* entries and restore them   */
-  for(;(current-begin)<size;current+=(strlen(current)+1)){
+  for(;(size_t)(current-begin)<size;current+=(strlen(current)+1)){
     Environment = environ;             /* get the environment        */
     del = NULL;
     np = current;
