+++ kernel/runtime/Utilities.cpp
@@ -85,16 +85,16 @@ void report_nomethod(                  /* report a NOMETHOD condition       */
     report_exception2(Error_No_method_name, receiver, message);
 }
 
-long message_number(
+int message_number(
     RexxString *errorcode)             /* REXX error code as string         */
 /******************************************************************************/
 /* Function:  Parse out the error code string into the messagecode valuey     */
 /******************************************************************************/
 {
   const char *decimalPoint;            /* location of decimalPoint in errorcode*/
-  long primary;                        /* Primary part of error code, major */
-  long secondary;                      /* Secondary protion (minor code)    */
-  long count;
+  int  primary;                        /* Primary part of error code, major */
+  int  secondary;                      /* Secondary protion (minor code)    */
+  int  count;
 
                                        /* make sure we get errorcode as str */
   errorcode = (RexxString *)errorcode->stringValue();
@@ -105,7 +105,7 @@ long message_number(
   primary = (new_string(errorcode->getStringData(), count)->longValue(9)) * 1000;
                                        /* did major code compute to long    */
                                        /* and within range                  */
-  if (primary == NO_LONG || primary < 1 || primary >= 100000) {
+  if (primary == (int)NO_LONG || primary < 1 || primary >= 100000) {
                                        /* Nope raise an error.              */
     report_exception(Error_Expression_result_raise);
   }
@@ -114,7 +114,7 @@ long message_number(
                                        /* Yes, compute its decimal value.   */
     secondary = new_string(decimalPoint + 1, errorcode->getLength() - count -1)->longValue(9);
                                        /* is the subcode invalid or too big?*/
-    if (secondary == NO_LONG || secondary < 0  || secondary >= 1000) {
+    if (secondary == (int)NO_LONG || secondary < 0  || secondary >= 1000) {
                                        /* Yes, raise an error.              */
       report_exception(Error_Expression_result_raise);
     }
@@ -122,7 +122,7 @@ long message_number(
   else {
     secondary = 0;
   }
-  return primary + secondary;          /* add two protions together, return */
+  return primary + secondary;          /* add two portions together, return */
 }
 
 void process_new_args(
